backend/modules/requirements.txt
websockets==12.0
numpy==1.26.4
torch==2.2.2
vosk==0.3.45
# konlpy==0.6.0
pandas
matplotlib
opencv-python
deepface
Pillow
tf-keras